requirements.txt
streamlit==1.28.1
PyMuPDF==1.23.5
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
ibm-watson-machine-learning==1.0.335
ibm-cloud-sdk-core==3.16.7
transformers==4.33.2
torch==2.0.1
huggingface-hub==0.16.4
langchain==0.0.292
tiktoken==0.5.1